requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0

# Database and async support
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching (optional)
redis==5.0.1

# Performance monitoring
psutil==5.9.6

# Security and Authentication
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
cryptography==45.0.5
python-jose[cryptography]==3.3.0

# Security validation and monitoring
python-multipart==0.0.6
pydantic==2.5.0
email-validator==2.1.0
python-dateutil==2.8.2

# Rate limiting and security
slowapi==0.1.9

# HIPAA compliance and audit
structlog==23.2.0
